interface extraction
