+++ magicmetamodel/src/test/kotlin/org/jetbrains/magicmetamodel/MagicMetaModelTest.kt
@@ -1,7 +1,5 @@
 package org.jetbrains.magicmetamodel
 
-import org.junit.jupiter.api.BeforeEach
-
 
 class MagicMetaModelTest {
